+++ packages/core/src/index.ts
@@ -14,6 +14,7 @@ export * from "./co2";
 export * from "./csv";
 export * from "./csvParse";
 export * from "./damageClaims";
+export * from "./ddd";
 export * from "./driverSettlement";
 export * from "./duration";
 export * from "./enums";